refactored/extracted bash util
